Add files via upload
TIM'S AFTER EFFECTS SCRIPTS: Rectangles 1.3
This is a script for creating split-screen videos quickly. Create a rectangle, divide it up as you wish, and populate it with footage. You can depopulate populated rectangles, populate a rectangle with an existing footage layer, and quickly switch the locations of two or more clips.